+++ kko_okk/Views/HeaderViews/TitleView.swift
@@ -42,7 +42,7 @@ struct TitleBackgroundView: View {
 struct TitleLeadingView: View {
     //MARK: 이부분이 데이터 피커에서 선택된 날짜값을 가지고 있습니다.
     //코어 용사 구엘 화이팅!!
-    @State var date = Date()
+    @ObservedObject var pickedDate = PickedDate()
     //let tips: [TipModel] = TipModel.tips
     // 구엘이 미리 만들어 놓은 팁 모델에서 팁을 가져오는 것 입니다.
     var body: some View {
@@ -51,14 +51,14 @@ struct TitleLeadingView: View {
             
             ZStack{
                 HStack{
-                    Text("\(HeaderViewConst.shared.dateToString(date: date))")
+                    Text("\(HeaderViewConst.shared.dateToString(date: pickedDate.date))")
                         .font(Font.Kkookk.headerDate)
                         .foregroundColor(Color.Kkookk.commonBlack)
                     Spacer()
                 }
                 
                 HStack{
-                    DatePickerView(date: $date)
+                    DatePickerView(date: $pickedDate.date)
                         .frame(width: 20, height: 20, alignment: .leading)
                         .padding(.leading,220)
                     Spacer()
